Some changes
Some changes that would suit me and my guild. I added some names and abbreviations for the dungeons, I also added a few more things to the parameter descriptions. I added a validation to be able to send the time in two formats, 12 hours and 24 hours. I noticed that the schedule is not working 100% when we created the event, it worked a few times when I used 'now', but never when I used a date, I tried to correct it by always setting the parameter to 'now', when it received empty, but I don't believe it's the best solution,